Second (preliminary) version of Practicum 10th project (Sprint 10, Mesto (API 4 Ardrey 2 check), before 2nd code review

+++ src/scripts/components/PopupWithConfirmation.js
@@ -12,7 +12,7 @@ export default class PopupWithConfirmation extends PopupWithForm {
   }
 
   close(deleteCardHandler) {
-    deleteCardHandler(this.#cardID, this.#cardElement)
+    if(deleteCardHandler) deleteCardHandler(this.#cardID, this.#cardElement)
     super.close();
   }
 }
